+++ buchfink/models/__init__.py
@@ -1,3 +1,3 @@
 from .account import Account
 from .config import (AccountConfig, Config, ExchangeAccountConfig, FetchConfig,
-                     ManualAccountConfig, ReportConfig)
+                     HistoricalPriceConfig, ManualAccountConfig, ReportConfig)